splitted serving-api/driver
