--- driver/src/test/scala/com/stratio/sparkta/driver/test/LocalSparkContext.scala
@@ -1,43 +1,0 @@
-/**
- * Copyright (C) 2015 Stratio (http://stratio.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.stratio.sparkta.driver.test
-
-import org.apache.log4j.{Level, Logger}
-import org.apache.spark.SparkContext
-import org.scalatest.{BeforeAndAfterAll, FlatSpec}
-
-trait LocalSparkContext extends BeforeAndAfterAll {
-  self: FlatSpec =>
-
-  @transient var sc: SparkContext = _
-
-  override def beforeAll {
-    Logger.getRootLogger.setLevel(Level.ERROR)
-    sc = LocalSparkContext.getNewLocalSparkContext(1, "test")
-  }
-
-  override def afterAll {
-    sc.stop()
-    System.clearProperty("spark.driver.port")
-  }
-}
-
-object LocalSparkContext {
-
-  def getNewLocalSparkContext(numExecutors: Int = 1, title: String): SparkContext =
-    new SparkContext(s"local[$numExecutors]", title)
-}
