+++ serving-api/src/test/scala/com/stratio/sparkta/serving/api/test/StreamingContextServiceIT.scala
@@ -20,7 +20,7 @@ import java.io.File
 
 import akka.event.slf4j.SLF4JLogging
 import com.stratio.sparkta.driver.constants.AppConstant
-import com.stratio.sparkta.driver.helpers.sparkta.MockSystem
+import com.stratio.sparkta.driver.helpers.MockSystem
 import com.stratio.sparkta.driver.models.AggregationPoliciesModel
 import com.stratio.sparkta.driver.service.StreamingContextService
 import com.stratio.sparkta.sdk.JsoneyStringSerializer
